Huge progress. Now actually working.

+++ src/integer.h
@@ -11,7 +11,7 @@
 #ifndef __integer_h
 #define __integer_h
 
-double numeric_value( struct cons_pointer pointer );
+long double numeric_value( struct cons_pointer pointer );
 
 /**
  * Allocate an integer cell representing this value and return a cons pointer to it.
